requirements.txt
mock
nose
python-dateutil
kerberos
pep8
pylint==0.26.0
simplejson
oslo.config
ceilometer
git+git://github.com/Katello/katello-cli.git#egg=katello_cli

